Moving cloudnode to it's own module
